refactor: move file I/O from core to system (capability model)
slurp, spit, load-file and load-file-once move from core to the system
library. core becomes a pure base with no ambient filesystem access: it
keeps eval/read-program/read-string (evaluate code held as data) but can
no longer read a file to run it. File access is an explicit capability —
+system for raw I/O, +require for structured module loading (require
reads modules itself in Go, so it needs neither slurp nor system).

- Move slurp/slurp_source/spit + the VerifySource integrity hook from
  lib/core to lib/system; move the load-file header (header-load-file.lisp)
  and load-file-once from nscore.LoadInput to nssystem.Load. read-program
  and eval stay in core.
- command's setupIntegrity now installs system.VerifySource.
- Not user-facing: the lisp binary loads system by default, so scripts
  and CLI script execution (which uses load-file) are unaffected.
  Breaking for embedders loading only core — they must add nssystem.Load
  (after core + its input layer). Documented in the CHANGELOG.
- Updated the ~15 test harnesses that build core-only envs and use
  slurp/load-file to also load system/nssystem; docgen blurbs and
  LANGUAGE.md regenerated.
