+++ command/debug_release_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/jig/lisp"
 	"github.com/jig/lisp/env"
 	"github.com/jig/lisp/lib/core"
+	"github.com/jig/lisp/lib/system"
 )
 
 func TestExecute_DebugFlagRequiresDebugBuild(t *testing.T) {
@@ -20,7 +21,7 @@ func TestExecute_DebugFlagRequiresDebugBuild(t *testing.T) {
 	if _, err := lisp.REPL(ctx, ns, core.HeaderBasic(), nil); err != nil {
 		t.Fatalf("preamble failed: %v", err)
 	}
-	if _, err := lisp.REPL(ctx, ns, core.HeaderLoadFile(), nil); err != nil {
+	if _, err := lisp.REPL(ctx, ns, system.HeaderLoadFile(), nil); err != nil {
 		t.Fatalf("preamble failed: %v", err)
 	}
