+++ lib/core/core.go
@@ -17,7 +17,6 @@ import (
 	"fmt"
 	"math/big"
 	"os"
-	"path/filepath"
 	"runtime/debug"
 	"strings"
 	"sync"
@@ -33,7 +32,6 @@ import (
 	"github.com/jig/lisp/marshaler"
 	"github.com/jig/lisp/printer"
 	"github.com/jig/lisp/reader"
-	lispruntime "github.com/jig/lisp/runtime"
 
 	. "github.com/jig/lisp/types"
 )
@@ -41,11 +39,7 @@ import (
 //go:embed header-basic.lisp
 var headerBasic string
 
-//go:embed header-load-file.lisp
-var headerLoadFile string
-
-func HeaderBasic() string    { return headerBasic }
-func HeaderLoadFile() string { return headerLoadFile }
+func HeaderBasic() string { return headerBasic }
 
 func Load(env EnvType) {
 	call.Call(env, assoc_in)
@@ -298,9 +292,6 @@ func drop_last(n int, arg MalType) (MalType, error) {
 }
 
 func LoadInput(env EnvType) {
-	call.Call(env, slurp)
-	call.Call(env, slurp_source)
-	call.Call(env, spit, 2, 4)
 	call.Call(env, readLine)
 	call.CallOverrideFN(env, "read-password", readPassword)
 	call.Call(env, exit, 0, 1)
@@ -495,89 +486,6 @@ func println(a ...MalType) (MalType, error) {
 func printNoNewline(a ...MalType) (MalType, error) {
 	fmt.Print(printer.Pr_list(a, false, "", "", " "))
 	return nil, nil
-}
-
-// VerifySource is an optional hook that vets a source file before
-// load-file (via slurp-source) evaluates it; a non-nil error aborts the
-// load. The command package installs it when running under --integrity,
-// so code loaded at runtime is verified like the script and its
-// requires. slurp itself is never hooked: it reads data, not code.
-var VerifySource func(absPath string, content []byte) error
-
-// slurp_source is slurp for files that will be evaluated as code:
-// identical, except that under --integrity the content is verified
-// against the pinned commit. load-file builds on it.
-func slurp_source(fileName string) (MalType, error) {
-	v, err := slurp(fileName)
-	if err != nil {
-		return nil, err
-	}
-	if VerifySource != nil {
-		abs, err := filepath.Abs(fileName)
-		if err != nil {
-			return nil, err
-		}
-		if err := VerifySource(abs, []byte(v.(string))); err != nil {
-			return nil, err
-		}
-	}
-	return v, nil
-}
-
-func slurp(fileName string) (MalType, error) {
-	b, e := os.ReadFile(fileName)
-	if e != nil {
-		return nil, e
-	}
-	// Register the module so the debugger can map cursors from files
-	// loaded at runtime back to their on-disk source (load-file injects
-	// a `;; $MODULE <fileName>` prefix naming the module after this same
-	// path); without the mapping, stepping would skip those files as
-	// library code. Dead code in release builds.
-	if lispruntime.Enabled {
-		if abs, err := filepath.Abs(fileName); err == nil {
-			lispruntime.Modules.Register(fileName, abs)
-		}
-	}
-	return string(b), nil
-}
-
-// spit is the write counterpart of slurp, following Clojure's:
-// (spit filename s) creates or truncates the file, and
-// (spit filename s :append true) appends instead.
-func spit(fileName, contents string, opts ...MalType) error {
-	if len(opts)%2 != 0 {
-		return fmt.Errorf("spit: options must be keyword value pairs")
-	}
-	appendMode := false
-	for i := 0; i < len(opts); i += 2 {
-		switch opts[i] {
-		case NewKeyword("append"):
-			b, ok := opts[i+1].(bool)
-			if !ok {
-				return fmt.Errorf("spit: :append expects a boolean (it was %T)", opts[i+1])
-			}
-			appendMode = b
-		default:
-			return fmt.Errorf("spit: unknown option %s", printer.Pr_str(opts[i], true))
-		}
-	}
-	flags := os.O_WRONLY | os.O_CREATE
-	if appendMode {
-		flags |= os.O_APPEND
-	} else {
-		flags |= os.O_TRUNC
-	}
-	f, err := os.OpenFile(fileName, flags, 0o644)
-	if err != nil {
-		return err
-	}
-	_, werr := f.WriteString(contents)
-	cerr := f.Close()
-	if werr != nil {
-		return werr
-	}
-	return cerr
 }
 
 // osExit is the process-exit hook, indirected so tests can observe the
