+++ lib/system/header-load-file.lisp
@@ -5,6 +5,6 @@
     [file-path]
     (eval (read-program (slurp-source file-path) file-path)))
 
-;; load-file-once is registered in Go by nscore.LoadInput: its seen-set
+;; load-file-once is registered in Go by nssystem.Load: its seen-set
 ;; needs mutable state, and atoms live in the concurrent library, which
 ;; is not available at core-load time.
