+++ mal_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/jig/lisp/lib/concurrent"
 	"github.com/jig/lisp/lib/core"
 	"github.com/jig/lisp/lib/coreextended"
+	"github.com/jig/lisp/lib/system"
 	. "github.com/jig/lisp/types"
 )
 
@@ -33,7 +34,7 @@ func BenchmarkMAL1(b *testing.B) {
 		if _, err := REPL(ctx, repl_env, core.HeaderBasic(), NewCursorFile(b.Name())); err != nil {
 			b.Fatal(err)
 		}
-		if _, err := REPL(ctx, repl_env, core.HeaderLoadFile(), NewCursorFile(b.Name())); err != nil {
+		if _, err := REPL(ctx, repl_env, system.HeaderLoadFile(), NewCursorFile(b.Name())); err != nil {
 			b.Fatal(err)
 		}
 		if _, err := REPL(ctx, repl_env, coreextended.HeaderCoreExtended(), NewCursorFile(b.Name())); err != nil {
@@ -164,7 +165,7 @@ func TestAtomParallel(t *testing.T) {
 	if _, err := REPL(ctx, repl_env, core.HeaderBasic(), NewCursorFile(t.Name())); err != nil {
 		t.Fatal(err)
 	}
-	if _, err := REPL(ctx, repl_env, core.HeaderLoadFile(), NewCursorFile(t.Name())); err != nil {
+	if _, err := REPL(ctx, repl_env, system.HeaderLoadFile(), NewCursorFile(t.Name())); err != nil {
 		t.Fatal(err)
 	}
 	if _, err := REPL(ctx, repl_env, coreextended.HeaderCoreExtended(), NewCursorFile(t.Name())); err != nil {
@@ -213,7 +214,7 @@ func BenchmarkAtomParallel(b *testing.B) {
 	if _, err := REPL(ctx, repl_env, core.HeaderBasic(), NewCursorFile(b.Name())); err != nil {
 		b.Fatal(err)
 	}
-	if _, err := REPL(ctx, repl_env, core.HeaderLoadFile(), NewCursorFile(b.Name())); err != nil {
+	if _, err := REPL(ctx, repl_env, system.HeaderLoadFile(), NewCursorFile(b.Name())); err != nil {
 		b.Fatal(err)
 	}
 	if _, err := REPL(ctx, repl_env, coreextended.HeaderCoreExtended(), NewCursorFile(b.Name())); err != nil {
